Fix some style-related issue with the Markdown links and fields.
The Markdown reference links were just a bit too high up, which was
more noticeable now that we have a checkbox. This fixes them to be
better aligned.

Plain text fields were also getting a minimum height when they shouldn't
have, due to the way we checked for the presence of a minHeight option.
This resulted in the minHeight being set to the default of 70 when 0 was
passed. We now check for the existence by comparing to undefined.

Reviewed at https://reviews.reviewboard.org/r/6468/