Fix unit test failures with TextEditorView usage.
Some of the unit tests were breaking with the transition to
TextEditorView, in particular with the plain text backend. In some
cases, they were expecting access to methods that no longer existed, or
looking up the wrong data key.

This did expose a real bug in the change event handling, which now looks
for key events as well as change events, and allows the events to bubble
so that key bindings work.

Reviewed at https://reviews.reviewboard.org/r/6470/