Fix issues around the Whitespace Only logic and presentation.
If a file was processed that had no chunks, it would be listed as
whitespace-only. This was because the logic computing this would default
to whitespace-only, and then override that if it found any non-equal
chunks that contained real changes. For empty files, there were no
chunks to process, so this flag would be incorrectly set.

Even if the flag was correctly set, it still looked broken, as the
message describing that it's whitespace-only would be hidden by default.
The CSS rule for this has been removed.

While here, I also fixed up the labels for the toggles to not have such
varying mixed caps. That's been bugging me.

Reviewed at https://reviews.reviewboard.org/r/6485/