Compare file SHAs to determine if files can be skipped in interdiffs.
We attempt to keep unmodified changes out of the diff viewer when
viewing interdiffs. To do this, we compare the deleted flags and also
the diff contents. However, if the diff content simply has a small
change, like a difference in timestamps in the headers, this check will
fail and the otherwise unmodified file will be shown.

Now, when we generate diffs, we calculate SHAs and store them in the
file's extra_data. We can use these to compare the SHAs of patched files
to see if there will be any changes to show.

Reviewed at https://reviews.reviewboard.org/r/6486/