Add optional Markdown support for the review request page.
This hooks up the remaining bits of the UI on the review request page to
the richText settings. Banners, review request fields, and replies all
now properly load and save properly based on the richText settings.

Most of this is just passing in richText values around, or converting
them to text type values. There are also unit tests to check that the
proper values make it out to the API.

Reviewed at https://reviews.reviewboard.org/r/6487/