Fix selecting text within a single line in the diff viewer.
The new text selection code only worked if the selected text had a <tr>
in it. If it didn't, we'd end up with an empty clipboard.

We now check for this condition, and just grab the text contents of the
selection.

Fixes bug #3631
Reviewed at https://reviews.reviewboard.org/r/6492/