Fix warnings about naive datetimes.
When we added timezone support, we updated almost everything to use it. The one
thing that was missed was the widgets in the admin dashboard, since we didn't
look at them too much. This meant that the log files were getting spewed with
warnings any time a user viewed the dashboard.

Testing done:
Viewed the dashboard and saw that there were no more naive datetime warnings.

Reviewed at https://reviews.reviewboard.org/r/6497/