Only set a DB index for ReviewRequest.summary.
The previous change I had for adding an index for search added it in
BaseReviewRequestDetails. While we didn't need it for drafts, I figured
it didn't hurt.

However, it exposed a bug in django-evolution, where, presumably due to
the base class, applying the change to ReviewRequest's signature also
affected ReviewRequestDraft's.

I don't have a fix for that yet in django-evolution, so for now, I'm
splitting that field out and turning indexing on only for
ReviewRequest.summary.

Reviewed at https://reviews.reviewboard.org/r/6498/