Added alphabetic pagination to users list
The usergrid object triggers the paginator in djiblets to turn on alphabetic
queries by providing a current_letter parameter to the DataGrid. If there is no
letter parameter passed into query string it defaults to 'A'. The users_list
function (in views) is reponsible for catching a letter parameter in the query
string (including catching and throwing errors), and filtering the queryset
accordingly (ie. filter for names starting with 'A') before passing it to the
datagrid.

Testing Done:
- Manually tested button behaviour of first last, next and back arrows.
- Unit tests for the user list are updated to test new features and all pass.
- All unit tests pass.

Bugs closed: 829

Reviewed at https://reviews.reviewboard.org/r/6368/