Fix a JavaScript test discrepency between Firefox and Chrome.
On Firefox, the tests for reloading the New Review Request view failed,
due to a different CSS display value being used when showing the
branches view. Since Chrome and Firefox disagree here, and what we
really care about is that it's being shown again, we can just test that
it's not "display: none".

Reviewed at https://reviews.reviewboard.org/r/6534/