Fix loading plain text content after saving a field.
When saving a field as plain text, the new field content would be
returned as HTML (through the use of force_text_type=html). This would
result in certain characters being encoded (which is also true when
loading the page initially). We then attempted to encode the text on
load again, leading to double-encoding.

We now turn off the second encoding when we render the text, preventing
this from happening.

Reviewed at https://reviews.reviewboard.org/r/6542/