Fix issues with raw values passed to the page.
When storing data attributes for raw values for custom fields that
weren't yet set, we'd end up setting the value to "None," which would
appear when editing the field. This was clearly wrong. We now make sure
to use an empty string in this case.

For built-in fields, we were setting the data attributes for raw values
unnecessarily. We already have that state separately, so setting the
data attributes just led to wasted memory and bandwidth. We now make
sure not to send those along.

Reviewed at https://reviews.reviewboard.org/r/6548/