Relabel the per-user Markdown setting.
This changes the label to "Always use Markdown for text fields" in order
to more accurately describe the impact of the setting.

Previously, it claimed that new fields would just be in Markdown by
default, but really, this is about editing all fields in Markdown by
default, even if they're stored in the database as plain text.

Reviewed at https://reviews.reviewboard.org/r/6549/