Open the header field by default in the "Edit Review" dialog.
This change makes it so that the header field opens into an editor by default,
except in the case where there's some saved text in there already. This is done
because usually if someone is opening that dialog, it's to add some general
commentary. We might revisit this once the general comments work is done.

While I was in here, I noticed that the label for "Ship It" wasn't lining up
properly with its checkbox, because a CSS rule was too general and was making
it into a block element. I've fixed the rule to be more specific.

Testing done:
- Review dialog with no comments: open editor for the header.
- Review dialog with some comments: open editor for the header.
- Review dialog with pre-existing header text: closed editor.

Reviewed at https://reviews.reviewboard.org/r/6546/