Add ?include-text-types= to the API to replace ?include-raw-text-fields.
The new review dialog demonstrated a need to have not just HTML-forced
text in API payloads, but Markdown-forced as well. The API didn't
previously provide a way to make this happen.

This change introduces a new option for including multiple text types in
one payload. This directly replaces ?include-raw-text-fields, allowing
for the client to request raw text, HTML, Markdown, and/or plain text in
one request. These all end up in their own container fields in the
payload (such as markdown_text_fields, raw_text_fields, etc.). It also
works pretty well as a replacement for ?force-text-type= in most cases,
though does not necessarily warrant deprecating that option.

Documentation was updated for this support, with a note on the
deprecation of ?include-raw-text-fields=.

Reviewed at https://reviews.reviewboard.org/r/6553/