Stop accepting group display names in target_groups on a draft in the API.
The API for review request drafts was allowing target_groups to match up
group display names. This was here for old legacy reasons, but is
undocumented, and it's very unlikely anyone is making use of it anymore.
(If they are, they'll have to adapt.)

The main reason this was a problem was that a group could have a display
name matching another group's ID, which would result in an error in
adding that group. That case was confusing and non-obvious, and worth
making the above change.