Fix using the search field to jump to a review request by ID.
In Review Board 1.7.x, it was possible to enter an ID in the search
field and, if search was enabled for that installation, go directly to
the review request page.

The search rewrite in 2.0 lost this ability. This change adds it back
through the same mechanism.

We probably want to support this functionality directly in the quick
search widget in the future, so that it would work without having to
have full-on search enabled. In the meantime, however, this at least
fixes this particular regression.

It's also worth noting that in 1.7.x, this was broken with Local Sites,
as it tried to use database IDs and not display IDs. This is fixed
properly for 2.0.

Reviewed at https://reviews.reviewboard.org/r/6565/