Only show diff expansion links in review comments.
Add the `show_controls` parameter to `build_diff_fragments` which determines if
expansion controls should be shown for the diff fragments. This parameter
defaults to `False`.

The `comment_diff_fragments` view sets this `show_controls` parameter to `True`
if and only if it is showing the fragments in a review and not the edit review
	dialog.

Testing Done:
Tested on local server:

 - Review comments contain expansion links.
 - Edit review dialog does not contain expansion links.
 - Emails do not contain expansion links.

Reviewed at https://reviews.reviewboard.org/r/6567/