Sandbox reviewboard.accounts.AuthBackend
Poorly-written AuthBackend implementations provided by extensions could
theoretically break parts of Review Board. While a number of the AuthBackend
calls are done from Django (and therefore can't be changed), there are several
Review Board-specific features that can be wrapped with sandboxing.

Now when an AuthBackend subclass from an extension throws an exception, Review
Board logs the errors with enough information to find which method in the
AuthBackend subclass threw the exception.

Testing Done:
Unit tests have been written to make sure that functions from an AuthBackend
subclass have been called, and when an exception is thrown it gets logged.

The tests fail without the sanboxing, and succeed with it.

Reviewed at https://reviews.reviewboard.org/r/6399/