Remove an old comment about Repository path:local_site constraints.
We had an old comment describing how a path:local_site unique constraint
was causing problems when archiving repositories. This was an issue
because an archived repository would keep the path, but wouldn't be
shown in the UI, resulting in a conflict when trying to add a new
repository with the same path.

We fixed this in 2.0 by including an archive timestamp in the key. The
comment hasn't been relevant for a while now, so we can safely delete
it.