Fix some text-related issues in the new review dialog.
The new review dialog had some quirks with the new Markdown code.
Publishing would end up saving HTML-formatted versions of the text,
and saving then editing the body top/bottom text would result in an
empty field.

The HTML-formatting issue was due to the way in which we load and then
save the review. Upon loading, we'd force text to HTML, which would then
be saved back out when doing the publish. We had similar issues with the
Ship It and Open An Issue checkboxes.

This was fixed by using the new fine-grained saving, making sure to only
save the fields we care about. I also made some fixes to ensure that
toggling these checkboxes wouldn't cause the stored text to be displayed
as plain text, which was caused by not setting force-text-type and
include-text-types on all requests.

The "empty field on edit" issue was due to using an incorrect variable
when loading the raw value for a field after save. We now use the
correct value.

Reviewed at https://reviews.reviewboard.org/r/6586/