Fix loading review public states and resolving comment issues.
We were deserializing the 'public' state of reviews and replies
incorrectly, which led to problems when operating on any child
resources. One affected set of resources were the comment resources,
which became clearly broken when trying to resolve issues.

Once fixed, I noticed that, like the review dialog, closing an issue
would take the HTML-formatted text originally fetched and send that in
the request. We only need to send issueStatus, and don't want to
override the comment, so we now set an appropriate attrs option.

Reviewed at https://reviews.reviewboard.org/r/6601/