Don't attempt a doomed request when publishing review replies.
When publishing review replies, we'd attempt to reset the state. Part
of this operation was discarding the reply if empty. However, in the
case of publishing, we've just set up a new reply and know we don't have
a reply sitting around to reset.

We now simply skip this step if publishing.

This also fixes the banner persisting on the screen if discarding twice
in a row.

Reviewed at https://reviews.reviewboard.org/r/6603/