Drag 'n Drop Inline Images Backend - Model
Part of a set of changes to add support for dragging and dropping images into
the markdown editor.

This change set involves extending FileAttachment model so that it can be owned
by a user. This includes adding a user field and a local_site field to the
FileAttachment model.

Testing Done:
All unit tests pass after making the changes.

Reviewed at https://reviews.reviewboard.org/r/6614/